care_backend/requirements.txt
Babel==2.6.0
Django==2.2
django-bootstrap-modal-forms==1.4.2
django-rest-framework==0.1.0
django-seed==0.1.9
django-widget-tweaks==1.4.5
djangorestframework==3.9.2
drf-url-filters==0.5.1
Faker==2.0.2
Pillow==6.1.0
psycopg2==2.8.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
voluptuous==0.11.7
